+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @mondrian-framework/cli
 
+## 2.0.20
+
+### Patch Changes
+
+- Provider dependencies
+- Updated dependencies
+  - @mondrian-framework/ci-tools@2.0.20
+  - @mondrian-framework/cli-commander@2.0.20
+
 ## 2.0.19
 
 ### Patch Changes