+++ packages/module/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mondrian-framework/module
 
+## 2.0.20
+
+### Patch Changes
+
+- Provider dependencies
+- Updated dependencies
+  - @mondrian-framework/model@2.0.20
+
 ## 2.0.19
 
 ### Patch Changes